Avalanche: create algorithms module for non-trivial flattening
