+++ icicle.cabal
@@ -89,6 +89,7 @@ library
                        Icicle.Avalanche.Statement.Simp.Melt
                        Icicle.Avalanche.Statement.Simp.Mutate
                        Icicle.Avalanche.Statement.Flatten
+                       Icicle.Avalanche.Statement.Flatten.Algorithms
                        Icicle.Avalanche.Statement.Flatten.Base
                        Icicle.Avalanche.Statement.Flatten.Save
                        Icicle.Avalanche.Statement.Flatten.Statement
